Revert "Normalizando recebidos do api"
