--- src/normalizarTransacao.ts
@@ -1,43 +1,0 @@
-declare global {
-  type TransacaoPagamento = "Boleto" | "Cartão de Crédito";
-  type TransacaoStatus =
-    | "Recusada pela operadora de cartão"
-    | "Paga"
-    | "Aguardando pagamento"
-    | "Estornada";
-
-  interface TransacaoAPI {
-    Nome: string;
-    ID: number;
-    Data: string;
-    Status: TransacaoStatus;
-    Email: string;
-    ["Valor (R$)"]: string;
-    ["Forma de Pagamento"]: TransacaoPagamento;
-    ["Cliente Novo"]: number;
-  }
-  interface transacao {
-    nome: string;
-    id: number;
-    data: string;
-    status: TransacaoStatus;
-    email: string;
-    moeda: string;
-    valor: number | null;
-    pagamento: TransacaoPagamento;
-    novo: boolean;
-  }
-}
-export default function normalizarTransacao(transacao: TransacaoAPI) {
-  return {
-    nome: transacao.Nome,
-    id: transacao.ID,
-    data: transacao.Data,
-    status: transacao.Status,
-    email: transacao.Email,
-    moeda: transacao["Valor (R$)"],
-    valor: 0,
-    pagamento: transacao["Forma de Pagamento"],
-    novo: Boolean(transacao["Cliente Novo"]),
-  };
-}
