+++ src/script.ts
@@ -1,14 +1,33 @@
 import fetchData from "./fetchData.js";
-import normalizarTransacao from "./normalizarTransacao.js";
+
+type TransacaoPagamento = "Boleto" | "Cartão de Crédito";
+type TransacaoStatus =
+  | "Recusada pela operadora de cartão"
+  | "Paga"
+  | "Aguardando pagamento"
+  | "Estornada";
+
+interface TransacaoAPI {
+  Nome: string;
+  ID: number;
+  Data: string;
+  Status: TransacaoStatus;
+  Email: string;
+  ["Valor (R$)"]: string;
+  ["Forma de Pagamento"]: TransacaoPagamento;
+  ["Cliente Novo"]: number;
+}
 
 async function handleData() {
   const data = await fetchData<TransacaoAPI[]>(
     "https://api.origamid.dev/json/transacoes.json"
   );
-  if (!data) return;
 
-  const transacoes = data.map(normalizarTransacao);
-  console.log(transacoes);
+  if (data) {
+    data.forEach((item) => {
+      console.log(item.Nome);
+    });
+  }
 }
 
 handleData();
